🌱 Split compute.test.ts (1686 lines) into 3 focused test modules
Split by resource type as recommended in #10439 triage:

- compute.nodes.test.ts (~520 lines) — useNodes hook tests
- compute.gpu.test.ts (~1100 lines) — GPU nodes, cache, fetch, dedup
- compute.nvidia.test.ts (~280 lines) — NVIDIA operator tests

All 89 tests pass. Unused imports cleaned per file. No production code changed.

Addresses #10439
